requirements.txt
fastapi
uvicorn[standard]
pymongo
tensorflow
opencv-python-headless
numpy
pandas
Pillow
scipy
python-multipart
pydantic
